refactor: extract types to types.tsx
