+++ src/stories/button/types.tsx
@@ -1,0 +1,9 @@
+import { UiButtonProps } from './ui';
+
+export interface ButtonProps extends UiButtonProps {
+  primary?: boolean;
+  size?: 'small' | 'medium' | 'large';
+  label: string;
+  viewType?: 'default' | 'facebook' | 'google' | 'like' | 'login' | 'logout';
+  onClick?: () => void;
+}
