+++ src/stories/card-button/types.tsx
@@ -1,0 +1,3 @@
+export interface ShopButtonProps {
+  itemCount: number;
+}
