+++ src/stories/card-item/ui/types.tsx
@@ -1,0 +1,7 @@
+export interface ShopItemProps {
+  imgSrc: string;
+  name: string;
+  sku: number;
+  itemCount: number;
+  price: number;
+}
