+++ src/stories/categories/types.tsx
@@ -1,0 +1,9 @@
+export interface CategoryProps {
+  id: number;
+  name: string;
+}
+
+export interface CategoriesProps {
+  list: CategoryProps[] | [];
+  title: string;
+}
