+++ src/stories/full-description/types.tsx
@@ -1,0 +1,6 @@
+export interface FullDescriptionProps {
+  common: string;
+  living_room: string;
+  dinning_room: string;
+  office: string;
+}
