+++ src/stories/header/Header.tsx
@@ -4,19 +4,9 @@ import UiHeader from './ui';
 import Logo from '../logo/Logo';
 import NavList from './NavList';
 import { Button } from '../button/Button';
+import { HeaderProps } from './types';
 
-type User = {
-  name: string;
-};
-
-interface HeaderProps {
-  user?: User;
-  onLogin: () => void;
-  onLogout: () => void;
-  onCreateAccount: () => void;
-}
-
-export const Header = ({ user, onLogin, onLogout, onCreateAccount }: HeaderProps) => {
+export const Header = ({ user, onLogin, onLogout }: HeaderProps) => {
   return (
     <UiHeader>
       <UiHeader.Logo href="/">
