+++ src/stories/layout/Layout.tsx
@@ -1,10 +1,7 @@
-import React, { ReactNode } from 'react';
+import React from 'react';
 import UiLayout from './ui';
 import { Header } from '../header/Header';
-
-interface LayoutProps {
-  children: ReactNode;
-}
+import { LayoutProps } from './types';
 
 const Layout: React.FC<LayoutProps> = ({ children }) => {
   const onLogin = () => {
