+++ src/stories/layout/types.tsx
@@ -1,0 +1,5 @@
+import { ReactNode } from 'react';
+
+export interface LayoutProps {
+  children: ReactNode;
+}
