+++ src/stories/modal/types.tsx
@@ -1,0 +1,7 @@
+import { ReactNode } from 'react';
+import { UiModalProps } from './ui';
+
+export interface ModalProps extends UiModalProps {
+  visible: boolean;
+  children: ReactNode;
+}
