+++ src/stories/product-social/types.tsx
@@ -1,0 +1,9 @@
+export interface SocialItemProps {
+  id: string;
+  link: string;
+  viewType: 'email' | 'facebook' | 'linkedin' | 'twitter';
+}
+
+export interface SocialListProps {
+  list: SocialItemProps[] | [];
+}
