+++ src/stories/rating/types.tsx
@@ -1,0 +1,5 @@
+export interface RatingProps {
+  rating: number;
+  title: string;
+  reviewCount: number;
+}
