+++ src/stories/slider/types.tsx
@@ -1,0 +1,9 @@
+export interface Image {
+  id: number;
+  src: string;
+  alt: string;
+}
+
+export interface ProductSliderProps {
+  images: Image[];
+}
