+++ src/stories/tabs/types.tsx
@@ -1,0 +1,5 @@
+export interface TabData {
+  id: number;
+  name: string;
+  component: JSX.Element;
+}
